requirements.txt
pip==25.1.1
langchain
langchain-huggingface
chromadb
transformers
accelerate
sentence-transformers
fastapi
uvicorn
python-dotenv
pydantic
pytest
gradio